move S3Uploader into its own file
